detector/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
ultralytics>=8.0.0 --extra-index-url https://download.pytorch.org/whl/cpu

